refactor(B6): split 1263-line supabase.ts god file into domain modules
The monolithic supabase.ts contained auth, session management, toilet
CRUD, and review CRUD all in one file. It was the #1 recommended
refactor from the codebase audit (HIGH severity).

Split into 5 focused modules:
- supabaseClient.ts (66 lines) — singleton client + env validation
- supabaseSession.ts (245 lines) — refresh, health checks, timestamps
- supabaseAuth.ts (516 lines) — auth operations + profile CRUD
- supabaseToilets.ts (133 lines) — toilet getNearby/getById/create
- supabaseReviews.ts (189 lines) — review CRUD

supabase.ts is now a 33-line barrel re-export preserving the original
export shape, so zero consumer imports needed updating.
